Episode 23

+++ app/Http/Controllers/ThreadsController.php
@@ -111,6 +111,12 @@ class ThreadsController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
+        // if ($thread->user_id != auth()->id()) {
+
+        //     abort(403, 'You do not have permission to do this.');
+
+        // }
+        $this->authorize('update', $thread);
 
         $thread->delete();
